backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
# Fine-tuning dependencies
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
bitsandbytes>=0.41.0
datasets>=2.14.0
accelerate>=0.24.0
